Add files via upload
This is a TODO app which is used to notify us with our daily activities to be performed. It has an input box where you can enter your daily tasks and add it to your TODO List. We can also update a task and also can delete task from entered list. It also has an extra feature as Clear List which clears all the tasks from the list.